Refactor favorite action
