+++ src/actions/changeNameAction.js
@@ -1,7 +1,7 @@
 
 export const CHANGE_NAME = 'CHANGE_NAME';
 
-export default (name) => {
+export default name => {
   return {
     type: CHANGE_NAME,
     name
