+++ src/actions/favoriteActions.js
@@ -1,6 +1,6 @@
 export const UPDATE_FAVORITE = 'UPDATE_FAVORITE';
 
-export default (value, favoriteType) => (
+export const update = (value, favoriteType) => (
   {
     type: UPDATE_FAVORITE,
     favoriteType,
